use different file to store models
